+++ src/game/blocks/BlockModifier.java
@@ -1,17 +1,18 @@
 package game.blocks;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 public interface BlockModifier {
-    default public void onDig(Vector3f pos) throws Exception {
+    default public void onDig(Vector3d pos) throws Exception {
 //        System.out.println("digging interface worked");
     }
 
-    default public void onPlace(Vector3f pos) {
+    default public void onPlace(Vector3d pos) {
 
     }
 
-    default public void onRightClick(Vector3f pos){
+    default public void onRightClick(Vector3d pos){
 
     }
 }
