+++ src/game/item/ItemModifier.java
@@ -1,9 +1,10 @@
 package game.item;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 public interface ItemModifier {
-    default public void onPlace(Vector3f pos) {
+    default public void onPlace(Vector3d pos) {
         System.out.println("placing interface worked");
     }
